dirt-orbit: allow lock delay
Global effects can be linked to the current `cps` parameter, if set up
appropriately. This commit allows to beat lock the delay time, if the
`lock` parameter is set to 1. By default it is off. Intermediate values
interpolate between locked and unlocked. You can switch it on by
`~d1.defaultParentEvent[\lock] = 1`.