dirt: new panning scheme
This is a new panning scheme, where in multiple input channels are
distributed (“splayed”) across multiple output channels.

The conversion unipolar -> bipolar in the `pan` parameter is now done
in sclang, which is more efficient.

This is an API change. The defaultMixingFunction is replaced by a
defaultPanningFunction, so that the full panning behaviour can be
changed if needed.