Global effects with new behavior
Demonstrates a new behavior for global effects with two features
  * any sound can independently control the "send" amount to an effect
  * other events that don't use the global effect won't interfere with the effect's parameters

This should make it a little easier to work with multiple streams to a single orbit when global effects are used.

This first commit demonstrates the idea with two new global effects: a tape delay and a gated reverb.

Example use in Tidal:
```
d1
 $ s "bd <~ ~ ~ cr> sn:2 <~ ~ ~ cr>" # release 1 # cut 1
 # gateverb 4 # gateverbd 0.5 # gateverbg 0.3 # gateverbr 1
 # gain 0.8

d3
 $ superimpose (### [bandf "[400,800,1300]", bandq 5])
 $ s "arpy*8" # up ("<0 0 3 -5>") # shape 0.2 # gain 0.8
 # tape 0.5 # taped 1 # tapefb 0.3 |*| speed 4
```