dirt-event: default sample number is \none
Because Symbols are transparent to math operations (this fact is used
in sclang a lot) we can use a symbol as a default for sample number and
thereby know if any frequency related parameter (note, midinote, or
freq) was explicitly set. This helps in midi event to detect whether a
note on is intended or not.

This changes two behaviors:
- so far, one could omit note and just play with octaves. Now we have
to specify note to do this.
- now, the default value for frequency in synth events is taken from
the SynthDef (this is arguably an improvement) and doesn’t override it
with the default from the event.